refactor: move more core utlis to core
