+++ skeliner/__init__.py
@@ -1,8 +1,8 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from . import batch, dx, io, pair, plot, post
-from .dataclass import Skeleton
+from .dataclass import Skeleton, register_skeleton_methods
 from .skeletonize import skeletonize
+from . import batch, dx, io, pair, plot, post
 from .plot.vis2d import projection as plot2d
 from .plot.vis2d import threeviews as plot3v
 from .plot.vis3d import view3d
@@ -11,6 +11,9 @@ try:
     __version__ = version(__name__)       
 except PackageNotFoundError:              
     __version__ = "0.0.0.dev0"
+
+register_skeleton_methods(dx, getattr(dx, "__skeleton__", None))
+register_skeleton_methods(post, getattr(post, "__skeleton__", None))
 
 __all__ = [
     "Skeleton",
