+++ skeliner/skeletonize.py
@@ -9,7 +9,15 @@ import numpy as np
 import trimesh
 from scipy.spatial import KDTree
 
-from .core import _bfs_parents, _build_mst, _find_soma
+from .core import (
+    _bfs_parents,
+    _bridge_gaps,
+    _build_mst,
+    _estimate_radius,
+    _find_soma,
+    _merge_near_soma_nodes,
+    _merge_single_node_branches,
+)
 from .dataclass import Skeleton, Soma
 
 _SKELINER_VERSION = _metadata.version("skeliner")
@@ -283,36 +291,6 @@ def _bin_geodesic_shells(
     return all_shells
 
 
-def _estimate_radius(
-    d: np.ndarray,
-    *,
-    method: str = "median",
-    trim_fraction: float = 0.05,
-    q: float = 0.90,
-) -> float:
-    """Return one scalar radius according to *method*.
-
-    Methods
-    -------
-    median          : 50‑th percentile (robust default)
-    mean            : arithmetic mean (biased by tails)
-    trim            : mean after trimming *trim_fraction* at both ends
-    qXX / pXX       : upper quantile XX given as integer, e.g. "q90" or "p75"
-    """
-    if method == "median":
-        return float(np.median(d))
-    if method == "mean":
-        return float(d.mean())
-    if method == "max":
-        return float(d.max())
-    if method == "min":
-        return float(d.min())
-    if method == "trim":
-        lo, hi = np.quantile(d, [trim_fraction, 1.0 - trim_fraction])
-        return float(d[(d >= lo) & (d <= hi)].mean())
-    raise ValueError(f"Unknown radius estimator '{method}'.")
-
-
 def _edges_from_mesh(
     edges_unique: np.ndarray,  # (E, 2) int64
     v2n: dict[int, int],  # mesh-vertex id -> skeleton node id
@@ -337,394 +315,6 @@ def _edges_from_mesh(
     edges = np.sort(edges, axis=1)  # canonical order
     edges = np.unique(edges, axis=0)  # drop duplicates
     return edges.astype(np.int64)  # copy to new array
-
-
-# -----------------------------------------------------------------------
-#  Post-processing helpers
-# -----------------------------------------------------------------------
-
-
-def _merge_near_soma_nodes(
-    nodes: np.ndarray,
-    radii_dict: dict[str, np.ndarray],
-    edges: np.ndarray,
-    node2verts: list[np.ndarray],
-    *,
-    soma: Soma,
-    radius_key: str,
-    mesh_vertices: np.ndarray,
-    # ---- new, all relative to the fitted spherical radius -------------
-    inside_tol: float = 0.0,  # anything with d < 0 − tol is *inside*
-    near_factor: float = 1.2,  # “near” if d < near_factor × r_soma
-    fat_factor: float = 0.20,  # “fat” if r ≥ fat_factor  × r_soma
-    log: Callable | None = None,
-):
-    """
-    Collapse every skeleton node whose sphere would overlap the soma
-    **according to strictly geometric tests based on `Soma.distance_`.**
-
-    *Inside test*
-        `d < −inside_tol`  (negative ➜ center is strictly inside)
-
-    *Near-and-fat test*
-        `d <  near_factor × r_soma   AND   r ≥ fat_factor × r_soma`
-
-    All factors are *dimensionless* and therefore unit-safe.
-    """
-    # ------------------------------------------------------------------
-    r_soma = soma.spherical_radius
-    r_primary = radii_dict[radius_key]
-    d2s = soma.distance(nodes, to="surface")  # signed distance
-    d2c = soma.distance(nodes, to="center")  # unsigned distance
-    inside = d2s < -inside_tol
-    near = d2c < near_factor * r_soma
-    fat = r_primary > fat_factor * r_soma
-
-    keep_mask = ~(inside | (near & fat))
-    keep_mask[0] = True  # never drop the soma vertex
-    merged_idx = np.where(~keep_mask)[0]
-
-    if log:
-        log(f"{merged_idx.size} nodes merged into soma")
-
-    # ---- remainder is identical to your original routine -------------
-    old2new = -np.ones(len(nodes), np.int64)
-    old2new[np.where(keep_mask)[0]] = np.arange(keep_mask.sum())
-
-    a, b = edges.T
-    both_keep = keep_mask[a] & keep_mask[b]
-    edges_out = np.unique(np.sort(old2new[edges[both_keep]], 1), axis=0)
-
-    leaves = {
-        int(old2new[u if keep_mask[u] else v])
-        for u, v in edges
-        if keep_mask[u] ^ keep_mask[v]
-    }
-    if leaves:
-        edges_out = np.vstack(
-            [edges_out, np.array([[0, i] for i in sorted(leaves)], dtype=np.int64)]
-        )
-
-    nodes_keep = nodes[keep_mask].copy()
-    radii_keep = {k: v[keep_mask].copy() for k, v in radii_dict.items()}
-    node2_keep = [node2verts[i] for i in np.where(keep_mask)[0]]
-    vert2node = {}
-
-    if merged_idx.size:
-        w = np.array(
-            [len(node2verts[0]), *[len(node2verts[i]) for i in merged_idx]],
-            dtype=np.float64,
-        )
-        nodes_keep[0] = np.average(
-            np.vstack([nodes[0], nodes[merged_idx]]), axis=0, weights=w
-        )
-        for idx in merged_idx:
-            # keep only the vertex subset that is *really* close to the soma
-            vidx = node2verts[idx]
-            d_local = soma.distance_to_surface(mesh_vertices[vidx])
-            close = d_local < near_factor * r_soma  # same factor as for center test
-            if np.any(close):
-                soma.verts = (
-                    np.concatenate((soma.verts, vidx[close]))
-                    if soma.verts is not None
-                    else vidx[close]
-                )
-                node2_keep[0] = np.concatenate((node2_keep[0], vidx[close]))
-
-    for nid, verts in enumerate(node2_keep):
-        for v in verts:
-            vert2node[int(v)] = nid
-
-    soma.verts = (
-        np.unique(soma.verts).astype(np.int64) if soma.verts is not None else None
-    )
-    try:
-        soma = Soma.fit(mesh_vertices[soma.verts], verts=soma.verts)
-    except ValueError:
-        if log:
-            log("Soma fitting failed, using spherical approximation instead.")
-        # fallback to spherical approximation
-        soma = Soma.from_sphere(soma.center, soma.spherical_radius, verts=soma.verts)
-
-    nodes_keep[0] = soma.center
-    r_soma = soma.spherical_radius
-    for k in radii_keep:
-        radii_keep[k][0] = r_soma
-
-    if log:
-        centre_txt = ", ".join(f"{c:7.1f}" for c in soma.center)
-        radii_txt = ",".join(f"{c:7.1f}" for c in soma.axes)
-        log(f"Moved soma to [{centre_txt}]")
-        log(f"(r = {radii_txt})")
-
-    return (nodes_keep, radii_keep, node2_keep, vert2node, edges_out, soma)
-
-
-def _bridge_gaps(
-    nodes: np.ndarray,
-    edges: np.ndarray,
-    *,
-    bridge_max_factor: float | None = None,
-    bridge_recalc_after: int | None = None,
-) -> np.ndarray:
-    """
-    Bridge all disconnected surface components of a neuron mesh **back to the
-    soma component** by inserting synthetic edges.
-
-    The routine works in four logical stages:
-
-    1.  **Component analysis** – build an undirected graph of the mesh,
-        identify connected components, and mark the one that contains the
-        soma (vertex 0) as the *island*.
-    2.  **Gap prioritisation** – for every *foreign* component find the
-        geodesically closest vertex pair (component ↔ island) and push the
-        tuple ``(gap_distance, cid, idx_comp, idx_island)`` into a
-        min-heap.
-        If *bridge_max_factor* is *None* we estimate a conservative upper
-        bound from the initial gap distribution:
-
-        ``factor = clip( 55-th percentile(gaps) / ⟨edge⟩ , [6 ×, 12 ×] )``
-
-        This filters out pathologically long jumps right from the start.
-    3.  **Greedy growth** – repeatedly pop the nearest component from the
-        heap and connect it with **one** synthetic edge (the cached closest
-        pair).  After each merge the island KD-tree is rebuilt and the heap
-        entries are refreshed every *bridge_recalc_after* merges
-        (auto-chosen if *None*; ≈ 5 % of remaining gaps, capped at 32).
-        A stall counter and a gentle *relax_factor* (1.5) guarantee
-        termination even on meshes with extremely uneven gap sizes.
-    4.  **Finish** – return the original edges plus all new bridges,
-        sorted and de-duplicated.
-
-    Notes
-    -----
-    * Only **one** edge per foreign component is added; the global MST step
-      later will prune any redundant cycles that could arise.
-    * Complexity is dominated by KD-tree queries:
-      *O((|V_island| + Σ|V_comp|) log |V|)* in practice.
-    * The heuristic defaults trade a few hundred ms of runtime for a markedly
-      lower rate of “long-jump” bridges.  Power users can override
-      *bridge_max_factor* or *bridge_recalc_after* if desired.
-
-    Parameters
-    ----------
-    nodes
-        ``(N, 3)`` float64 array of mesh-vertex coordinates.
-    edges
-        ``(E, 2)`` int64 array of **undirected, sorted** mesh edges.
-    bridge_max_factor
-        Optional hard ceiling for acceptable bridge length expressed as a
-        multiple of the mean mesh-edge length.  If *None* an adaptive value
-        (see above) is chosen.
-    bridge_recalc_after
-        How many successful merges to perform before all component-to-island
-        distances are recomputed.  If *None* an adaptive value based on the
-        number of gaps is used.
-
-    Returns
-    -------
-    np.ndarray
-        ``(E′, 2)`` int64 undirected edge list containing the original mesh
-        edges **plus** one synthetic edge for every formerly disconnected
-        component.  The array is sorted row-wise and de-duplicated.
-    """
-
-    def _auto_bridge_max(
-        gaps: list[float],
-        edge_mean: float,
-        *,
-        pct: float = 55.0,
-        lo: float = 6.0,
-        hi: float = 12.0,
-    ) -> float:
-        """
-        Choose a bridge_max_factor from the initial gap distribution. Default is the
-        55th percentile of the gap distribution, clipped to [6, 20] times the mean edge
-        """
-        raw = np.percentile(gaps, pct) / edge_mean
-        return float(np.clip(raw, lo, hi))
-
-    def _auto_recalc_after(n_gaps: int) -> int:
-        """Return a suitable recalc period for the given number of gaps."""
-        if n_gaps <= 10:  # tiny: update often
-            return 2
-        if n_gaps <= 50:  # small: every 3–4 merges
-            return 4
-        if n_gaps <= 200:  # medium: every ~5 % of gaps
-            return max(4, n_gaps // 20)
-        # giant meshes: cap to 32 so we never starve
-        return 32
-
-    # -- 0. quick exit if already connected ---------------------------------
-    g = ig.Graph(
-        n=len(nodes), edges=[tuple(map(int, e)) for e in edges], directed=False
-    )
-    comps = [set(c) for c in g.components()]
-    comp_idx = {
-        cid: np.fromiter(verts, dtype=np.int64) for cid, verts in enumerate(comps)
-    }
-    soma_cid = g.components().membership[0]
-    if len(comps) == 1:
-        return edges
-
-    # -- 1. build one KD-tree per component ---------------------------------
-    edge_len_mean = np.linalg.norm(
-        nodes[edges[:, 0]] - nodes[edges[:, 1]], axis=1
-    ).mean()
-
-    # -- 2. grow the island using a distance-ordered priority queue ---------
-    island = set(comps[soma_cid])
-    island_idx = np.fromiter(island, dtype=np.int64)
-    island_tree = KDTree(nodes[island_idx])
-
-    # helper to compute the *current* closest gap of a component
-    def closest_pair(cid: int) -> tuple[float, int, int]:
-        pts = nodes[comp_idx[cid]]
-        dists, idx_is = island_tree.query(pts, k=1, workers=-1)
-        best = int(np.argmin(dists))
-        return float(dists[best]), best, np.asarray(idx_is, dtype=np.int64)[best]
-
-    # priority queue of (gap_distance, cid, best_comp_idx, best_island_idx)
-    pq = []
-    gap_samples = []
-    for cid in range(len(comps)):
-        if cid == soma_cid:
-            continue
-        gap, b_comp, b_is = closest_pair(cid)
-        pq.append((gap, cid, b_comp, b_is))
-        gap_samples.append(gap)
-    heapq.heapify(pq)
-
-    # -- heuristic hyperparameters if not given -------------------------------
-    if bridge_max_factor is None:
-        bridge_max_factor = _auto_bridge_max(gap_samples, edge_len_mean)
-
-    if bridge_recalc_after is None:
-        gaps = len(comps) - 1
-        recalc_after = _auto_recalc_after(gaps)
-    else:
-        recalc_after = bridge_recalc_after
-
-    edges_new: list[tuple[int, int]] = []
-    merges_since_recalc = 0
-
-    stall = 0
-    relax_factor = 1.5
-    max_stall = 3 * len(pq)
-    current_max = bridge_max_factor * edge_len_mean
-
-    while pq:
-        _, cid, _, _ = heapq.heappop(pq)
-        # postpone if the component is still too far away
-        gap, best_c, best_i = closest_pair(cid)
-
-        if gap > current_max:
-            heapq.heappush(pq, (gap, cid, best_c, best_i))  # still too far, re-queue
-            stall += 1
-            if stall >= 2 * max_stall:
-                # after too many futile tries, do a *forced* heap rebuild
-                pq = [
-                    (g, cid2, bc, bi)
-                    for _, cid2, _, _ in pq
-                    for g, bc, bi in (closest_pair(cid2),)
-                ]
-                heapq.heapify(pq)
-                current_max *= relax_factor
-                stall = 0
-            continue
-
-        stall = 0
-        verts_idx = comp_idx[cid]
-        u = int(verts_idx[best_c])
-        v = int(island_idx[best_i])
-        edges_new.append((u, v))
-
-        # merge component into island and rebuild KD-tree
-        island |= comps[cid]
-        island_idx = np.concatenate([island_idx, verts_idx])
-        island_tree = KDTree(nodes[island_idx])
-
-        merges_since_recalc += 1
-        if merges_since_recalc >= recalc_after:
-            # distances of *every* remaining component may have changed
-            pq = [
-                (gap, cid, b_comp, b_is)
-                for _, cid, _, _ in pq
-                for gap, b_comp, b_is in (closest_pair(cid),)
-            ]
-            heapq.heapify(pq)
-            merges_since_recalc = 0
-
-    if edges_new:
-        edges_aug = np.vstack([edges, np.asarray(edges_new, dtype=np.int64)])
-    else:
-        edges_aug = edges
-
-    return np.unique(edges_aug, axis=0)
-
-
-def _merge_single_node_branches(
-    nodes: np.ndarray,
-    radii_dict: dict[str, np.ndarray],
-    node2verts: list[np.ndarray],
-    edges: np.ndarray,
-    *,
-    mesh_vertices: np.ndarray,
-    min_parent_degree: int = 3,
-) -> tuple[
-    np.ndarray, dict[str, np.ndarray], list[np.ndarray], dict[int, int], np.ndarray
-]:
-    """
-    Iteratively merge every leaf whose *parent* has degree ≥ `min_parent_degree`.
-    Terminates when no such leaves are left.
-    """
-    while True:
-        # --- current degree -------------------------------------------
-        deg = np.zeros(len(nodes), dtype=int)
-        for a, b in edges:
-            deg[a] += 1
-            deg[b] += 1
-
-        # --- leaves that qualify --------------------------------------
-        leaves = [i for i, d in enumerate(deg) if d == 1 and i != 0]
-        parent = _bfs_parents(edges, len(nodes), root=0)
-        singles = [leaf for leaf in leaves if deg[parent[leaf]] >= min_parent_degree]
-
-        if not singles:
-            break  # fixed-point reached – nothing more to merge
-
-        # --- merge all singles in *one* shot ---------------------------
-        to_drop = set()
-        for leaf in singles:
-            par = parent[leaf]
-
-            # move vertices to parent
-            node2verts[par] = np.concatenate((node2verts[par], node2verts[leaf]))
-
-            # re-fit radii
-            pts = mesh_vertices[node2verts[par]]
-            d = np.linalg.norm(pts - nodes[par], axis=1)
-            for k in radii_dict:
-                radii_dict[k][par] = _estimate_radius(d, method=k)
-
-            to_drop.add(leaf)
-
-        # rebuild arrays after this sweep
-        keep = np.ones(len(nodes), bool)
-        keep[list(to_drop)] = False
-        remap = {old: new for new, old in enumerate(np.where(keep)[0])}
-
-        nodes = nodes[keep]
-        node2verts = [node2verts[i] for i in np.where(keep)[0]]
-        radii_dict = {k: v[keep].copy() for k, v in radii_dict.items()}
-        edges = np.asarray(
-            [(remap[a], remap[b]) for a, b in edges if keep[a] and keep[b]],
-            dtype=np.int64,
-        )
-
-    # final vert-to-node map
-    vert2node = {int(v): i for i, vs in enumerate(node2verts) for v in vs}
-    return nodes, radii_dict, node2verts, vert2node, edges
 
 
 def _prune_neurites(
