+++ tests/test_post.py
@@ -12,7 +12,7 @@ import numpy as np
 import pytest
 
 from skeliner import dx, post, skeletonize
-from skeliner.core import Skeleton, Soma
+from skeliner.dataclass import Skeleton, Soma
 from skeliner.io import load_mesh
